Extend the group node to its component gene nodes

Extend the group node to its component gene nodes

Usage

process.groupNode(groupNode, components, g, edge.attrs)

Arguments

groupNode

The group node to be processed

components

The vector of its component gene nodes

g

The pathway graph of type GraphNEL

edge.attrs

The list of edge attr

Details

The all the edges to and from the group node are